requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-ninja==1.1.0
gunicorn==22.0.0
packaging==24.0
pydantic==2.7.2
pydantic_core==2.18.3
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
